pyproject.toml
[tool.poetry]
name = "apiconf-agent"
version = "0.1.0"
description = "AI Agent for API Conference Community Support"
authors = ["Claret greyisheep@gmail.com"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.12"
uvicorn = {extras = ["standard"], version = "^0.34.2"}
google-adk = "^1.2.1"
google-generativeai = "^0.4.1"
googlemaps = "^4.10.0"
beautifulsoup4 = "^4.12.3"
requests = "^2.32.3"
pydantic = "^2.11.3"
pydantic-settings = "^2.9.1"
python-dotenv = "^1.0.0"
aiohttp = "^3.9.0"
pandas = "^2.2.2"
python-dateutil = "^2.8.2"
geopy = "^2.4.0"
redis = "^5.2.1"
sqlalchemy = "^2.0.40"
asyncpg = "^0.30.0"
alembic = "^1.15.2"
httpx = "^0.28.1"
lxml = "^5.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.0.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
mypy = "^1.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true 